pyproject.toml
[project]
name = "azure-multimodal-compilance-ingestion-engine"
version = "2.0.0"
description = "Production-grade multi-agent multimodal compliance auditing pipeline powered by Azure AI and LangGraph."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Azure SDK
    "azure-identity>=1.25.2",
    "azure-monitor-opentelemetry>=1.8.6",
    "azure-search-documents>=11.6.0",
    "azure-storage-blob>=12.28.0",
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "gunicorn>=23.0.0",
    "python-multipart>=0.0.20",
    # Rate limiting
    "slowapi>=0.1.9",
    "redis>=5.2.0",
    # LangChain / LangGraph / LangSmith
    "langchain>=0.3.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=0.3.0",
    "langgraph>=0.4.0",
    "langsmith>=0.3.0",
    # Tavily web-search (agentic tool)
    "tavily-python>=0.5.0",
    # Utilities
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "pypdf>=6.7.5",
    "yt-dlp>=2026.3.3",
    # Async performance
    "uvloop>=0.21.0",
    # Observability
    "opentelemetry-instrumentation-fastapi>=0.53b0",
    "azure-monitor-opentelemetry>=1.6.4",
]

